# Improve FineFundamental backtesting source resolution performance