--- a/Engine/DataFeeds/Enumerators/Factories/FineFundamentalSubscriptionEnumeratorFactory.cs
+++ b/Engine/DataFeeds/Enumerators/Factories/FineFundamentalSubscriptionEnumeratorFactory.cs
@@ -15,6 +15,7 @@
 */
 
 using System;
+using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Globalization;
 using System.IO;
@@ -33,6 +34,9 @@ namespace QuantConnect.Lean.Engine.DataFeeds.Enumerators.Factories
     /// </summary>
     public class FineFundamentalSubscriptionEnumeratorFactory : ISubscriptionEnumeratorFactory
     {
+        private static readonly ConcurrentDictionary<int, List<DateTime>> FineFilesCache
+            = new ConcurrentDictionary<int, List<DateTime>>();
+
         private readonly bool _isLiveMode;
         private readonly Func<SubscriptionRequest, IEnumerable<DateTime>> _tradableDaysProvider;
         private string _lastUsedFileName;
@@ -101,47 +105,107 @@ namespace QuantConnect.Lean.Engine.DataFeeds.Enumerators.Factories
 
             if (!File.Exists(source.Source))
             {
-                if (_lastUsedFileName == null)
+                // only use cache in backtest, since in live mode new fine files are added
+                if (!_isLiveMode)
                 {
-                    // find first file date
-                    var path = Path.GetDirectoryName(source.Source) ?? string.Empty;
-                    if (string.IsNullOrEmpty(path) || !Directory.Exists(path))
-                        return source;
-
-                    var firstFileName = Path.GetFileNameWithoutExtension(Directory.GetFiles(path, "*.zip").OrderBy(x => x).First());
-                    var firstDate = DateTime.ParseExact(firstFileName, "yyyyMMdd", CultureInfo.InvariantCulture);
-
-                    // requested date before first date, return current invalid source anyway
-                    if (date < firstDate)
-                        return source;
-
-                    // requested date after first date, save date of first existing file
-                    _lastUsedFileName = firstFileName;
-
-                    // loop back in time until we find an existing file
-                    while (string.CompareOrdinal(fileName, _lastUsedFileName) > 0)
+                    var cacheKey = config.Symbol.Value.ToLower().GetHashCode();
+                    List<DateTime> availableDates;
+                    // we still didn't load available fine dates for this symbol
+                    if (!FineFilesCache.TryGetValue(cacheKey, out availableDates))
                     {
-                        // get previous date
-                        date = date.AddDays(-1);
+                        try
+                        {
+                            var path = Path.GetDirectoryName(source.Source) ?? string.Empty;
+                            availableDates = Directory.GetFiles(path, "*.zip")
+                                .Select(
+                                    filePath =>
+                                    {
+                                        try
+                                        {
+                                            return DateTime.ParseExact(
+                                                Path.GetFileNameWithoutExtension(filePath),
+                                                "yyyyMMdd",
+                                                CultureInfo.InvariantCulture
+                                            );
+                                        }
+                                        catch
+                                        {
+                                            // just in case...
+                                            return DateTime.MaxValue;
+                                        }
+                                    }
+                                )
+                                .Where(time => time != DateTime.MaxValue)
+                                .OrderBy(x => x)
+                                .ToList();
+                        }
+                        catch
+                        {
+                            // directory doesn't exist or path is null
+                            FineFilesCache[cacheKey] = new List<DateTime>();
+                            return source;
+                        }
+                        FineFilesCache[cacheKey] = availableDates;
+                    }
 
-                        // get file name for this date
-                        source = fine.GetSource(config, date, _isLiveMode);
-                        fileName = Path.GetFileNameWithoutExtension(source.Source);
-
-                        if (!File.Exists(source.Source))
-                            continue;
-
-                        // we found the file, save its name and return the source
-                        _lastUsedFileName = fileName;
-
-                        break;
+                    // requested date before first date, return null source
+                    if (availableDates.Count == 0 || date < availableDates[0])
+                    {
+                        return source;
+                    }
+                    for (var i = availableDates.Count - 1; i >= 0; i--)
+                    {
+                        // we iterate backwards ^ and find the first data point before 'date'
+                        if (availableDates[i] <= date)
+                        {
+                            return fine.GetSource(config, availableDates[i], _isLiveMode);
+                        }
                     }
                 }
                 else
                 {
-                    // return source for last existing file date
-                    date = DateTime.ParseExact(_lastUsedFileName, "yyyyMMdd", CultureInfo.InvariantCulture);
-                    source = fine.GetSource(config, date, _isLiveMode);
+                    if (_lastUsedFileName == null)
+                    {
+                        // find first file date
+                        var path = Path.GetDirectoryName(source.Source) ?? string.Empty;
+                        if (string.IsNullOrEmpty(path) || !Directory.Exists(path))
+                            return source;
+
+                        var firstFileName = Path.GetFileNameWithoutExtension(Directory.GetFiles(path, "*.zip").OrderBy(x => x).First());
+                        var firstDate = DateTime.ParseExact(firstFileName, "yyyyMMdd", CultureInfo.InvariantCulture);
+
+                        // requested date before first date, return current invalid source anyway
+                        if (date < firstDate)
+                            return source;
+
+                        // requested date after first date, save date of first existing file
+                        _lastUsedFileName = firstFileName;
+
+                        // loop back in time until we find an existing file
+                        while (string.CompareOrdinal(fileName, _lastUsedFileName) > 0)
+                        {
+                            // get previous date
+                            date = date.AddDays(-1);
+
+                            // get file name for this date
+                            source = fine.GetSource(config, date, _isLiveMode);
+                            fileName = Path.GetFileNameWithoutExtension(source.Source);
+
+                            if (!File.Exists(source.Source))
+                                continue;
+
+                            // we found the file, save its name and return the source
+                            _lastUsedFileName = fileName;
+
+                            break;
+                        }
+                    }
+                    else
+                    {
+                        // return source for last existing file date
+                        date = DateTime.ParseExact(_lastUsedFileName, "yyyyMMdd", CultureInfo.InvariantCulture);
+                        source = fine.GetSource(config, date, _isLiveMode);
+                    }
                 }
             }
             else