--- a/Tests/Engine/DataFeeds/Enumerators/Factories/FineFundamentalSubscriptionEnumeratorFactoryTests.cs
+++ b/Tests/Engine/DataFeeds/Enumerators/Factories/FineFundamentalSubscriptionEnumeratorFactoryTests.cs
@@ -34,7 +34,8 @@ namespace QuantConnect.Tests.Engine.DataFeeds.Enumerators.Factories
     [TestFixture]
     public class FineFundamentalSubscriptionEnumeratorFactoryTests
     {
-        [Test, TestCaseSource(nameof(GetFineFundamentalTestParameters))]
+        [TestCaseSource(nameof(GetFineFundamentalTestParametersBacktest))]
+        [TestCaseSource(nameof(GetFineFundamentalTestParametersLive))]
         public void ReadsFineFundamental(FineFundamentalTestParameters parameters)
         {
             var stopwatch = Stopwatch.StartNew();
@@ -158,7 +159,13 @@ namespace QuantConnect.Tests.Engine.DataFeeds.Enumerators.Factories
             Assert.IsTrue(ramUsageAfterLoop - ramUsageBeforeLoop < 10);
         }
 
-        private static TestCaseData[] GetFineFundamentalTestParameters()
+        private static TestCaseData[] GetFineFundamentalTestParametersBacktest =>
+            GetFineFundamentalTestParameters(false);
+
+        private static TestCaseData[] GetFineFundamentalTestParametersLive =>
+            GetFineFundamentalTestParameters(true);
+
+        private static TestCaseData[] GetFineFundamentalTestParameters(bool liveMode)
         {
             return new List<FineFundamentalTestParameters>
             {
@@ -167,14 +174,16 @@ namespace QuantConnect.Tests.Engine.DataFeeds.Enumerators.Factories
                     Symbol = Symbols.AAPL,
                     StartDate = new DateTime(2014, 1, 1),
                     EndDate = new DateTime(2014, 12, 31),
-                    RowCount = 365
+                    RowCount = 365,
+                    LiveMode = liveMode
                 },
                 new FineFundamentalTestParameters("AAPL-BeforeFirstDate")
                 {
                     Symbol = Symbols.AAPL,
                     StartDate = new DateTime(2014, 2, 20),
                     EndDate = new DateTime(2014, 2, 20),
-                    RowCount = 1
+                    RowCount = 1,
+                    LiveMode = liveMode
                 },
                 new FineFundamentalTestParameters("AAPL-FirstDate")
                 {
@@ -188,7 +197,8 @@ namespace QuantConnect.Tests.Engine.DataFeeds.Enumerators.Factories
                     CostOfRevenue3M = 35748000000m,
                     CostOfRevenue12M = 106606000000m,
                     EquityPerShareGrowth1Y = 0.091652m,
-                    PeRatio = 13.012858m
+                    PeRatio = 13.012858m,
+                    LiveMode = liveMode
                 },
                 new FineFundamentalTestParameters("AAPL-BeforeLastDate")
                 {
@@ -202,7 +212,8 @@ namespace QuantConnect.Tests.Engine.DataFeeds.Enumerators.Factories
                     CostOfRevenue3M = 35748000000m,
                     CostOfRevenue12M = 106606000000m,
                     EquityPerShareGrowth1Y = 0.091652m,
-                    PeRatio = 13.272502m
+                    PeRatio = 13.272502m,
+                    LiveMode = liveMode
                 },
                 new FineFundamentalTestParameters("AAPL-LastDate")
                 {
@@ -216,7 +227,8 @@ namespace QuantConnect.Tests.Engine.DataFeeds.Enumerators.Factories
                     CostOfRevenue3M = 27699000000m,
                     CostOfRevenue12M = 106606000000m,
                     EquityPerShareGrowth1Y = 0.091652m,
-                    PeRatio = 13.272502m
+                    PeRatio = 13.272502m,
+                    LiveMode = liveMode
                 },
                 new FineFundamentalTestParameters("AAPL-AfterLastDate")
                 {
@@ -230,7 +242,8 @@ namespace QuantConnect.Tests.Engine.DataFeeds.Enumerators.Factories
                     CostOfRevenue3M = 27699000000m,
                     CostOfRevenue12M = 106606000000m,
                     EquityPerShareGrowth1Y = 0.091652m,
-                    PeRatio = 13.272502m
+                    PeRatio = 13.272502m,
+                    LiveMode = liveMode
                 },
 
             }.Select(x => new TestCaseData(x).SetName(x.Name)).ToArray();
@@ -239,6 +252,7 @@ namespace QuantConnect.Tests.Engine.DataFeeds.Enumerators.Factories
         public class FineFundamentalTestParameters
         {
             public string Name { get; }
+            public bool LiveMode { get; set; }
             public Symbol Symbol { get; set; }
             public DateTime StartDate { get; set; }
             public DateTime EndDate { get; set; }