<!--- Provide a general summary of your changes in the Title above -->

#### Description
<!--- Describe your changes in detail -->
- Reduce the amount of `Path.Combine()` usages -> it has a performance
overhead
- For backtesting, adding new private static fine cache that will hold the available dates
for which there is a fine file, reducing the number of times we
enumerate the files in the directory. Updating unit tests.
- `DefaultDataProvider` will not check if file exists since `new
FileStream` performs the same operation internally

`Cloud performance tests CoarseFineFundamentalComboAlgorithm 2015-2016 - 100 coarse and 10 fine symbols`
`.OrderByDescending(x => x.DollarVolume)`
`pr`
103.86 seconds at 0k data points per second. Processing total of 28,167 data points.
95.29 seconds at 0k data points per second. Processing total of 28,168 data points.
93.20 seconds at 0k data points per second. Processing total of 28,168 data points.

`master`
109.59 seconds at 0k data points per second. Processing total of 28,168 data points.
119.64 seconds at 0k data points per second. Processing total of 28,167 data points.
123.88 seconds at 0k data points per second. Processing total of 28,167 data points.

`.OrderBy(x => x.DollarVolume)` -> **these symbols are most likely to have less fine files**
`pr`
234.42 seconds at 0k data points per second. Processing total of 28,688 data points.
217.90 seconds at 0k data points per second. Processing total of 28,688 data points.
`master`
405.12 seconds at 0k data points per second. Processing total of 28,688 data points.
436.87 seconds at 0k data points per second. Processing total of 28,688 data points.

#### Related Issue
<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps to reproduce -->
<!--- Please link to the issue here: -->
Closes https://github.com/QuantConnect/Lean/issues/3403
#### Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
- Improve FineFundamental backtesting performance
#### Requires Documentation Change
<!--- Please indicate if these changes will require updates to documentation, and if so, specify what changes are required -->
N/A
#### How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->
Unit and regression tests, cloud testing
#### Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] Non-functional change (xml comments/documentation/etc)

#### Checklist:
<!--- The following is a checklist of items that MUST be completed before a PR is accepted -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the code style of this project.
- [x] I have read the **CONTRIBUTING** [document](https://github.com/QuantConnect/Lean/blob/master/CONTRIBUTING.md).
- [x] I have added tests to cover my changes. <!--- If not applicable, please explain why -->
- [x] All new and existing tests passed.
- [x] My branch follows the naming convention `bug-<issue#>-<description>` or `feature-<issue#>-<description>`

<!--- Template inspired by https://www.talater.com/open-source-templates/#/page/99 -->