@jaredbroad 2019-07-23
Closing in favor of caching at `TextSubscriptionDataSourceReader` level.